requirements.txt
# ── Web framework ──────────────────────────────────────────
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ── Database ───────────────────────────────────────────────
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# ── Auth ───────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.1.0

# ── Config ─────────────────────────────────────────────────
pydantic-settings==2.7.1

# ── RAG: vector store & embeddings ─────────────────────────
qdrant-client>=1.10,<2
fastembed>=0.7,<1
rank-bm25>=0.2,<0.3

# ── RAG: document readers ─────────────────────────────────
pymupdf>=1.24,<2
pypdf>=5,<6
python-docx>=1.1,<2

# ── RAG: LLM & HTTP ───────────────────────────────────────
openai>=1.0,<2
httpx>=0.27,<1

# ── Presentation ──────────────────────────────────────────
python-pptx>=1.0,<2

# ── Table extraction ─────────────────────────────────────
openpyxl>=3.1,<4

# ── Web Parser ───────────────────────────────────────────
beautifulsoup4>=4.12,<5
trafilatura>=1.6,<2
readability-lxml>=0.8,<1

# ── Audio / TTS (Silero) ──────────────────────────────────
torch>=2.1.0
silero
numpy>=1.24.0
omegaconf>=2.3.0

# ── Image (slides for video/presentation) ────────────────
Pillow>=10.0.0

# ── Testing & Linting ───────────────────────────────────
pytest>=8.0
ruff>=0.4
